feat(tests): add crontab helper and update test binaries in Makefile
