+++ lock/lock_test.go
@@ -12,6 +12,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/creativeprojects/resticprofile/platform"
 	"github.com/creativeprojects/resticprofile/shell"
 	"github.com/shirou/gopsutil/v3/process"
 	"github.com/stretchr/testify/assert"
@@ -26,7 +27,7 @@ func TestMain(m *testing.M) {
 	// using an anonymous function to handle defer statements before os.Exit()
 	exitCode := func() int {
 		var err error
-		lockBinary = filepath.Join(os.Getenv("TEST_HELPERS"), "test-args")
+		lockBinary = filepath.Join(os.Getenv("TEST_HELPERS"), platform.Executable("test-args"))
 		lockBinary, err = filepath.Abs(lockBinary)
 		if err != nil {
 			fmt.Fprintf(os.Stderr, "Failed to get absolute path of test-args binary: %v\n", err)
