+++ testhelpers/generate_windows.go
@@ -1,0 +1,7 @@
+//go:build windows
+
+package testhelpers
+
+//go:generate go build -o ../build/test-args.exe -buildvcs=false ./args
+//go:generate go build -o ../build/test-echo.exe -buildvcs=false ./echo
+//go:generate go build -o ../build/test-crontab.exe -buildvcs=false ./crontab
